Cleaned up macro frame info code
