+++ external-crates/move/crates/move-compiler/src/to_bytecode/mod.rs
@@ -5,5 +5,6 @@
 mod canonicalize_handles;
 #[macro_use]
 mod context;
+mod macro_frame_diagnostics;
 mod optimize;
 pub mod translate;
